pyproject.toml
[project]
name = "ai-newsletter-curator"
version = "0.1.0"
description = "Personal knowledge management system for AI newsletters"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Steve Merritt" }
]

dependencies = [
    # Email ingestion
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "imap-tools>=1.5.0",
    
    # HTML parsing and content extraction
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "html2text>=2020.1.16",
    "httpx>=0.25.0",  # For fetching linked articles
    "trafilatura>=1.6.0",  # Article content extraction from web pages
    
    # Embeddings and vector storage
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",  # Token counting for chunking
    # Uncomment for local embeddings (no OpenAI dependency):
    # "sentence-transformers>=2.2.0",
    
    # Data modeling and validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    
    # UI
    "streamlit>=1.29.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "rich>=13.0.0",  # Nice CLI output
    "tenacity>=8.2.0",  # Retry logic for API calls
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ipython>=8.0.0",
]

local-embeddings = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

[project.scripts]
ingest = "scripts.run_ingestion:main"
curator = "src.app.streamlit_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | \.mypy_cache
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
